requirements.txt
africastalking==1.2.9
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
flask==3.1.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.5
markupsafe==3.0.2
mysql-connector-python==9.2.0 
pip==25.0.1
pymysql==1.1.1 
python-dotenv==1.0.1
requests==2.32.3
schema==0.7.7
setuptools==65.5.0
sqlalchemy==2.0.38
typing-extensions==4.12.2
urllib3==2.3.0
uv==0.5.29
werkzeug==3.1.3
